Remove NetSnmpCmd base class
