+++ LibreNMS/OS/Axos.php
@@ -76,7 +76,7 @@ class Axos extends OS implements OSDiscovery, TransceiverDiscovery
         $inventory = new Collection;
         $physical_index = 1;
 
-        $physical_name = NetSnmpTranslate::make->hideMib()->mibs(['CALIX-PRODUCT-MIB'])->translate($this->getDevice()->sysObjectID);
+        $physical_name = NetSnmpTranslate::make()->hideMib()->mibs(['CALIX-PRODUCT-MIB'])->translate($this->getDevice()->sysObjectID);
         $serial_number = \SnmpQuery::get('Axos-System-MIB::axosSystemChassisSerialNumber.0')->value();
         $inventory->push(new EntPhysical([
             'entPhysicalIndex' => $physical_index++,
